requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
transformers>=4.20.0
clip-by-openai
open3d>=0.15.0
scikit-learn>=1.0.0
tqdm>=4.64.0
matplotlib>=3.5.0
tensorboard>=2.8.0
pyyaml>=6.0
imageio>=2.19.0
